requirements.txt
ccxt>=4.1.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
TA-Lib>=0.4.28
